requirements.txt
pyyaml>=6.0
websocket-client>=1.6
